backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
# Optional for deep learning model:
# tensorflow==2.17.0